ByteNCrunch/requirements.txt
anyio==3.7.1
APScheduler==3.6.3
blinker==1.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.2
decorator==5.1.1
Flask==3.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.2.0
protobuf==3.20.3
pycparser==2.21
python-dotenv==1.0.0
python-telegram-bot==13.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.1.0
Werkzeug==3.0.1
